完善增删改查

+++ src/main/java/com/metazz/metazzspace/service/impl/CategoryServiceImpl.java
@@ -6,12 +6,15 @@ import com.baomidou.mybatisplus.core.conditions.query.LambdaQueryWrapper;
 import com.baomidou.mybatisplus.extension.plugins.pagination.Page;
 import com.baomidou.mybatisplus.extension.service.impl.ServiceImpl;
 import com.metazz.metazzspace.common.exception.BaseException;
+import com.metazz.metazzspace.model.dto.BlogQueryDTO;
 import com.metazz.metazzspace.model.dto.CategoryAddDTO;
 import com.metazz.metazzspace.model.dto.CategoryModifyDTO;
 import com.metazz.metazzspace.model.dto.PageQueryDTO;
+import com.metazz.metazzspace.model.entity.Blog;
 import com.metazz.metazzspace.model.entity.Category;
 import com.metazz.metazzspace.common.enums.ExceptionEnum;
 import com.metazz.metazzspace.mapper.CategoryMapper;
+import com.metazz.metazzspace.service.IBlogService;
 import com.metazz.metazzspace.service.ICategoryService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -23,6 +26,9 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
 
     @Autowired
     CategoryMapper categoryMapper;
+
+    @Autowired
+    IBlogService blogService;
 
     @Override
     public Page<Category> getAllCategory(PageQueryDTO pageQueryDTO) {
@@ -46,6 +52,11 @@ public class CategoryServiceImpl extends ServiceImpl<CategoryMapper, Category> i
         Category category = categoryMapper.selectById(id);
         if(!Optional.ofNullable(category).isPresent()){
             throw new BaseException(ExceptionEnum.CATEGORY_NOT_EXISTS);
+        }
+        // 该分类关联了博客，不允许删除
+        Page<Blog> blogByPage = blogService.getBlogByPage(BlogQueryDTO.builder().current(1).size(5).categoryId(Integer.valueOf(id)).labelId(0).build());
+        if(CollectionUtil.isNotEmpty(blogByPage.getRecords())){
+            throw new BaseException(ExceptionEnum.CATEGORY_LINKED_BLOG);
         }
         categoryMapper.deleteById(id);
     }
